rd-agent: Embed misc support bins into rd-agent binary
rd-agent already does the same thing with sideloader bins. Generalize the
mechanism and embed iocost_coef_gen.py and sideloader.py into rd-agent binary.
