+++ install.sh
@@ -16,9 +16,7 @@ echo "[ Creating target/resctl-demo.tar.gz ]"
 tar cvzf target/resctl-demo.tar.gz --transform 's|^.*/\([^/]*\)$|\1|' \
     target/release/rd-hashd \
     target/release/rd-agent \
-    target/release/resctl-demo \
-    misc/iocost_coef_gen.py \
-    misc/sideloader.py
+    target/release/resctl-demo
 
 if [ -n "$1" ]; then
     echo "[ Installing under $1 ]"
