+++ rd-agent/src/misc.rs
@@ -1,0 +1,18 @@
+// Copyright (c) Facebook, Inc. and its affiliates.
+use super::{prepare_bin_file, Config};
+use anyhow::Result;
+
+const MISC_BINS: [(&str, &[u8]); 2] = [
+    (
+        "iocost_coef_gen.py",
+        include_bytes!("misc/iocost_coef_gen.py"),
+    ),
+    ("sideloader.py", include_bytes!("misc/sideloader.py")),
+];
+
+pub fn prepare_misc_bins(cfg: &Config) -> Result<()> {
+    for (name, body) in &MISC_BINS {
+        prepare_bin_file(&format!("{}/{}", &cfg.misc_bin_path, name), body)?;
+    }
+    Ok(())
+}
